refactor fcos
