+++ easycv/models/detection/utils/__init__.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
 # Copyright (c) 2014-2021 Megvii Inc. All rights reserved.
-
-from .boxes import (bbox2result, bboxes_iou, box_cxcywh_to_xyxy,
-                    box_xyxy_to_cxcywh, generalized_box_iou, postprocess, bbox_overlaps, distance2bbox)
-from .misc import (accuracy, interpolate, multi_apply, select_single_mlvl, filter_scores_and_topk, output_postprocess, fp16_clamp)
+from .boxes import (bbox2result, bbox_overlaps, bboxes_iou, box_cxcywh_to_xyxy,
+                    box_xyxy_to_cxcywh, distance2bbox, generalized_box_iou,
+                    postprocess)
+from .generator import MlvlPointGenerator
+from .misc import (accuracy, filter_scores_and_topk, fp16_clamp, interpolate,
+                   multi_apply, output_postprocess, select_single_mlvl)
